# docs/swagger.yaml
definitions:
  models.BannerMessage:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      expires_at:
        type: string
      id:
        type: string
      message:
        type: string
      scope_type:
        $ref: '#/definitions/models.ScopeType'
      scope_value:
        type: string
      severity:
        $ref: '#/definitions/models.Severity'
      starts_at:
        type: string
      updated_at:
        type: string
    type: object
...
  models.ScopeType:
    enum:
    - platform
    - service
    - page
    type: string
    x-enum-varnames:
    - ScopePlatform
    - ScopeService
    - ScopePage
  models.Severity:
    enum:
    - info
    - warning
    - critical
    type: string
    x-enum-varnames:
    - SeverityInfo
    - SeverityWarning
    - SeverityCritical
  server.createBannerRequest:
    properties:
      created_by:
        type: string
      expires_at:
        type: string
      message:
        type: string
      scope_type:
        $ref: '#/definitions/models.ScopeType'
      scope_value:
        type: string
      severity:
        $ref: '#/definitions/models.Severity'
      starts_at:
        type: string
    required:
    - created_by
    - expires_at
    - message
    - scope_type
    - severity
    type: object
...
  server.updateBannerRequest:
    properties:
      expires_at:
        type: string
      message:
        type: string
      scope_type:
        $ref: '#/definitions/models.ScopeType'
      scope_value:
        type: string
      severity:
        $ref: '#/definitions/models.Severity'
      starts_at:
        type: string
    required:
    - expires_at
    - message
    - scope_type
    - severity
    type: object
  vo.ErrorVO:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
info:
  contact:
    email: admin@sweetrpg.com
    name: API Support
    url: https://sweetrpg.com
  description: Swagger APIs
  license:
    name: MIT
    url: https://mit-license.org/
  termsOfService: https://pilgrimagesoftware.com/terms/
  title: Admin API service
  version: "1.0"
paths:
  /banners:
    get:
      description: Default mode returns active (started, not expired) banners matching
        any of the requested scopes, most-severe-first; platform-scoped banners are
        always included. scope values look like "platform", "service:catalog", or
        "page:/catalog/browse". Pass include_inactive=true (admin-web's management
        view) to instead return every banner regardless of scope or expiration - scope
        params are ignored in that mode.
      parameters:
      - collectionFormat: csv
        description: Repeated scope filter, e.g. scope=platform&scope=service:catalog
          - required unless include_inactive=true
        in: query
        items:
          type: string
        name: scope
        type: array
      - description: Return every banner (any scope, expired or scheduled included)
          for admin management views
        in: query
        name: include_inactive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BannerMessage'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorVO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorVO'
      summary: List banners
      tags:
      - banners
    post:
      consumes:
      - application/json
      description: Creates a new admin banner message. expires_at is required; a banner
        without one is rejected. Requires a forwarded user bearer token carrying the
        admin role.
      parameters:
      - description: Bearer user access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Banner message
        in: body
        name: banner
        required: true
        schema:
          $ref: '#/definitions/server.createBannerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BannerMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorVO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorVO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorVO'
      summary: Create a banner message
      tags:
      - banners
  /banners/{id}:
    delete:
      description: Deletes a banner message; it is immediately excluded from subsequent
        queries. Requires a forwarded user bearer token carrying the admin role.
      parameters:
      - description: Bearer user access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Banner ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorVO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorVO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/vo.ErrorVO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorVO'
      summary: Delete a banner message
      tags:
      - banners
    put:
      consumes:
      - application/json
      description: Replaces the mutable fields of an existing banner message. expires_at
        is required. Requires a forwarded user bearer token carrying the admin role.
      parameters:
      - description: Bearer user access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Banner ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated banner message
        in: body
        name: banner
        required: true
        schema:
          $ref: '#/definitions/server.updateBannerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BannerMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorVO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorVO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/vo.ErrorVO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vo.ErrorVO'
      summary: Update a banner message
      tags:
      - banners
...
swagger: "2.0"